[Chore] Move useSetMyBagList to Hook directory
